Extract work queue to allow reusing it
(cherry picked from commit 26862bb834b90722d6290181f8776711d8a7d10d)
